+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 2.x branch
 ## 2.31 branch
+### 2.31.1
+* fixed little bug for `ControllerTestCase::assert*Authorized()` methods;
+* added `AppController::getQueryPage()` method. Little fixes for controllers;
+* updated for php-tools 1.7.1.
+
 ### 2.31.0
 * uses `Authentication` and `Authorization` plugins instead of the old `AuthComponent`;
 * some controller code has been moved from the `beforeFilter()` method to `beforeRender()`;